prop group row to new fie
